Move the MarkerTooltipContents component to a better directory
